+++ skbio/stats/tests/test_subsample.py
@@ -18,6 +18,8 @@ import warnings
 
 import numpy as np
 
+from skbio.stats import subsample_items
+
 
 cy_subsample = import_fresh_module('skbio.stats._subsample',
                                    fresh=['skbio.stats.__subsample'])
@@ -132,6 +134,106 @@ class PySubsampleTests(SubsampleTests, unittest.TestCase):
                  "Accelerated subsample module unavailable.")
 class CySubsampleTests(SubsampleTests, unittest.TestCase):
     module = cy_subsample
+
+
+class SubsampleItemsTests(unittest.TestCase):
+    def setUp(self):
+        np.random.seed(123)
+
+        # comment indicates the expected random value
+        self.sequences = [
+            ('a_1', 'AATTGGCC-a1'),  # 2, 3624216819017203053
+            ('a_2', 'AATTGGCC-a2'),  # 5, 5278339153051796802
+            ('b_1', 'AATTGGCC-b1'),  # 4, 4184670734919783522
+            ('b_2', 'AATTGGCC-b2'),  # 0, 946590342492863505
+            ('a_4', 'AATTGGCC-a4'),  # 3, 4048487933969823850
+            ('a_3', 'AATTGGCC-a3'),  # 7, 7804936597957240377
+            ('c_1', 'AATTGGCC-c1'),  # 8, 8868534167180302049
+            ('a_5', 'AATTGGCC-a5'),  # 1, 3409506807702804593
+            ('c_2', 'AATTGGCC-c2'),  # 9, 8871627813779918895
+            ('c_3', 'AATTGGCC-c3')   # 6, 7233291490207274528
+        ]
+
+    def mock_sequence_iter(self, items):
+        return ({'SequenceID': sid, 'Sequence': seq} for sid, seq in items)
+
+    def test_subsample_items_simple(self):
+        maximum = 10
+        bin_f = lambda x: x['SequenceID'].rsplit('_', 1)[0]
+
+        # note, the result here is sorted by sequence_id but is in heap order
+        # by the random values associated to each sequence
+        exp = sorted([('a', {'SequenceID': 'a_5', 'Sequence': 'AATTGGCC-a5'}),
+                      ('a', {'SequenceID': 'a_1', 'Sequence': 'AATTGGCC-a1'}),
+                      ('a', {'SequenceID': 'a_4', 'Sequence': 'AATTGGCC-a4'}),
+                      ('a', {'SequenceID': 'a_3', 'Sequence': 'AATTGGCC-a3'}),
+                      ('a', {'SequenceID': 'a_2', 'Sequence': 'AATTGGCC-a2'}),
+                      ('b', {'SequenceID': 'b_2', 'Sequence': 'AATTGGCC-b2'}),
+                      ('b', {'SequenceID': 'b_1', 'Sequence': 'AATTGGCC-b1'}),
+                      ('c', {'SequenceID': 'c_3', 'Sequence': 'AATTGGCC-c3'}),
+                      ('c', {'SequenceID': 'c_2', 'Sequence': 'AATTGGCC-c2'}),
+                      ('c', {'SequenceID': 'c_1', 'Sequence': 'AATTGGCC-c1'})],
+                     key=lambda x: x[0])
+        obs = subsample_items(self.mock_sequence_iter(self.sequences), maximum,
+                              bin_f=bin_f)
+        self.assertEqual(sorted(obs, key=lambda x: x[0]), exp)
+
+    def test_per_sample_sequences_min_seqs(self):
+        maximum = 10
+        minimum = 3
+        bin_f = lambda x: x['SequenceID'].rsplit('_', 1)[0]
+
+        # note, the result here is sorted by sequence_id but is in heap order
+        # by the random values associated to each sequence
+        exp = sorted([('a', {'SequenceID': 'a_5', 'Sequence': 'AATTGGCC-a5'}),
+                      ('a', {'SequenceID': 'a_1', 'Sequence': 'AATTGGCC-a1'}),
+                      ('a', {'SequenceID': 'a_4', 'Sequence': 'AATTGGCC-a4'}),
+                      ('a', {'SequenceID': 'a_3', 'Sequence': 'AATTGGCC-a3'}),
+                      ('a', {'SequenceID': 'a_2', 'Sequence': 'AATTGGCC-a2'}),
+                      ('c', {'SequenceID': 'c_3', 'Sequence': 'AATTGGCC-c3'}),
+                      ('c', {'SequenceID': 'c_2', 'Sequence': 'AATTGGCC-c2'}),
+                      ('c', {'SequenceID': 'c_1', 'Sequence': 'AATTGGCC-c1'})],
+                     key=lambda x: x[0])
+        obs = subsample_items(self.mock_sequence_iter(self.sequences), maximum,
+                              minimum, bin_f=bin_f)
+        self.assertEqual(sorted(obs, key=lambda x: x[0]), exp)
+
+    def test_per_sample_sequences_complex(self):
+        maximum = 2
+        bin_f = lambda x: x['SequenceID'].rsplit('_', 1)[0]
+        exp = sorted([('a', {'SequenceID': 'a_2', 'Sequence': 'AATTGGCC-a2'}),
+                      ('a', {'SequenceID': 'a_3', 'Sequence': 'AATTGGCC-a3'}),
+                      ('b', {'SequenceID': 'b_2', 'Sequence': 'AATTGGCC-b2'}),
+                      ('b', {'SequenceID': 'b_1', 'Sequence': 'AATTGGCC-b1'}),
+                      ('c', {'SequenceID': 'c_1', 'Sequence': 'AATTGGCC-c1'}),
+                      ('c', {'SequenceID': 'c_2', 'Sequence': 'AATTGGCC-c2'})],
+                     key=lambda x: x[0])
+        obs = subsample_items(self.mock_sequence_iter(self.sequences), maximum,
+                              bin_f=bin_f, buf_size=1)
+        self.assertEqual(sorted(obs, key=lambda x: x[0]), exp)
+
+    def test_min_gt_max(self):
+        gen = subsample_items([1, 2, 3], maximum=2, minimum=10)
+        with self.assertRaises(ValueError):
+            next(gen)
+
+    def test_min_lt_zero(self):
+        gen = subsample_items([1, 2, 3], maximum=0, minimum=-10)
+        with self.assertRaises(ValueError):
+            next(gen)
+
+    def test_max_lt_zero(self):
+        gen = subsample_items([1, 2, 3], maximum=-10)
+        with self.assertRaises(ValueError):
+            next(gen)
+
+    def test_binf_is_none(self):
+        maximum = 2
+        items = [1, 2]
+        exp = [(True, 1), (True, 2)]
+        obs = subsample_items(items, maximum)
+        self.assertEqual(list(obs), exp)
+
 
 if __name__ == '__main__':
     import nose